first commit
This project started out without git.
This first commit is the first working version of the symfony2
chatroom webapp. Working version, does not mean fully and well
implemented. There is a huge ammount of work to be done. The data
structure implemented must be rapidly changed as well as the
javascript that's responsible for the client side update.